Fixes/game map (#5)
* item id as integer

* organize project

* organize project

* clean up pathfinder

* fix calc of gscore

* clean up pathfinder

* fix pickup

* room*tick just after loaded

* classSimpleName when register room process

* virtual id as bitwise flag

* check rights in incoming

* remove src folder

* feat: performance benchmarks

* apply code style

+++ api/src/main/java/habbo/rooms/components/objects/items/IRoomItemData.java
@@ -1,10 +1,10 @@
 package habbo.rooms.components.objects.items;
 
 import habbo.furniture.extra.data.IExtraData;
-import utils.Position;
+import utils.pathfinder.Position;
 
 public interface IRoomItemData {
-    long getId();
+    int getId();
 
     int getFurnitureId();
